42Quackform/requirements.txt
fastapi==0.110.2
jinja2==3.1.4
openai==1.30.1
pgvector==0.2.4
psycopg[binary]==3.1.18
python-dotenv==1.0.1
python-multipart==0.0.9
typer==0.9.0
uvicorn[standard]==0.29.0
spacy==3.7.4
